src/requirements.txt
requests
flask
python-dotenv
beautifulsoup4
